Migrate most users endpoints to the new pattern (#3366)
