+++ server/service/service_users.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/fleetdm/fleet/v4/server"
-	"github.com/fleetdm/fleet/v4/server/authz"
 
 	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
 	"github.com/fleetdm/fleet/v4/server/contexts/viewer"
@@ -40,27 +39,6 @@ func (svc *Service) CreateUserFromInvite(ctx context.Context, p fleet.UserPayloa
 		return nil, err
 	}
 	return user, nil
-}
-
-func (svc *Service) CreateUser(ctx context.Context, p fleet.UserPayload) (*fleet.User, error) {
-	var teams []fleet.UserTeam
-	if p.Teams != nil {
-		teams = *p.Teams
-	}
-	if err := svc.authz.Authorize(ctx, &fleet.User{Teams: teams}, fleet.ActionWrite); err != nil {
-		return nil, err
-	}
-
-	if invite, err := svc.ds.InviteByEmail(ctx, *p.Email); err == nil && invite != nil {
-		return nil, ctxerr.Errorf(ctx, "%s already invited", *p.Email)
-	}
-
-	if p.AdminForcedPasswordReset == nil {
-		// By default, force password reset for users created this way.
-		p.AdminForcedPasswordReset = ptr.Bool(true)
-	}
-
-	return svc.newUser(ctx, p)
 }
 
 func (svc *Service) CreateInitialUser(ctx context.Context, p fleet.UserPayload) (*fleet.User, error) {
@@ -106,157 +84,6 @@ func (svc *Service) newUser(ctx context.Context, p fleet.UserPayload) (*fleet.Us
 	return user, nil
 }
 
-func (svc *Service) ModifyUser(ctx context.Context, userID uint, p fleet.UserPayload) (*fleet.User, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.User{}, fleet.ActionRead); err != nil {
-		return nil, err
-	}
-
-	user, err := svc.User(ctx, userID)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := svc.authz.Authorize(ctx, user, fleet.ActionWrite); err != nil {
-		return nil, err
-	}
-
-	if p.GlobalRole != nil || p.Teams != nil {
-		if err := svc.authz.Authorize(ctx, user, fleet.ActionWriteRole); err != nil {
-			return nil, err
-		}
-	}
-	if p.Name != nil {
-		user.Name = *p.Name
-	}
-
-	if p.Email != nil && *p.Email != user.Email {
-		err = svc.modifyEmailAddress(ctx, user, *p.Email, p.Password)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if p.Position != nil {
-		user.Position = *p.Position
-	}
-
-	if p.GravatarURL != nil {
-		user.GravatarURL = *p.GravatarURL
-	}
-
-	if p.SSOEnabled != nil {
-		user.SSOEnabled = *p.SSOEnabled
-	}
-
-	currentUser := authz.UserFromContext(ctx)
-
-	if p.GlobalRole != nil && *p.GlobalRole != "" {
-		if currentUser.GlobalRole == nil {
-			return nil, ctxerr.New(ctx, "Cannot edit global role as a team member")
-		}
-
-		if p.Teams != nil && len(*p.Teams) > 0 {
-			return nil, fleet.NewInvalidArgumentError("teams", "may not be specified with global_role")
-		}
-		user.GlobalRole = p.GlobalRole
-		user.Teams = []fleet.UserTeam{}
-	} else if p.Teams != nil {
-		if !isAdminOfTheModifiedTeams(currentUser, user.Teams, *p.Teams) {
-			return nil, ctxerr.New(ctx, "Cannot modify teams in that way")
-		}
-		user.Teams = *p.Teams
-		user.GlobalRole = nil
-	}
-
-	err = svc.saveUser(ctx, user)
-	if err != nil {
-		return nil, err
-	}
-
-	return user, nil
-}
-
-func isAdminOfTheModifiedTeams(currentUser *fleet.User, originalUserTeams, newUserTeams []fleet.UserTeam) bool {
-	// If the user is of the right global role, then they can modify the teams
-	if currentUser.GlobalRole != nil && (*currentUser.GlobalRole == fleet.RoleAdmin || *currentUser.GlobalRole == fleet.RoleMaintainer) {
-		return true
-	}
-
-	// otherwise, gather the resulting teams
-	resultingTeams := make(map[uint]string)
-	for _, team := range newUserTeams {
-		resultingTeams[team.ID] = team.Role
-	}
-
-	// and see which ones were removed or changed from the original
-	teamsAffected := make(map[uint]struct{})
-	for _, team := range originalUserTeams {
-		if resultingTeams[team.ID] != team.Role {
-			teamsAffected[team.ID] = struct{}{}
-		}
-	}
-
-	// then gather the teams the current user is admin for
-	currentUserTeamAdmin := make(map[uint]struct{})
-	for _, team := range currentUser.Teams {
-		if team.Role == fleet.RoleAdmin {
-			currentUserTeamAdmin[team.ID] = struct{}{}
-		}
-	}
-
-	// and let's check that the teams that were either removed or changed are also teams this user is an admin of
-	for teamID := range teamsAffected {
-		if _, ok := currentUserTeamAdmin[teamID]; !ok {
-			return false
-		}
-	}
-
-	return true
-}
-
-func (svc *Service) modifyEmailAddress(ctx context.Context, user *fleet.User, email string, password *string) error {
-	// password requirement handled in validation middleware
-	if password != nil {
-		err := user.ValidatePassword(*password)
-		if err != nil {
-			return fleet.NewPermissionError("incorrect password")
-		}
-	}
-	random, err := server.GenerateRandomText(svc.config.App.TokenKeySize)
-	if err != nil {
-		return err
-	}
-	token := base64.URLEncoding.EncodeToString([]byte(random))
-	err = svc.ds.PendingEmailChange(ctx, user.ID, email, token)
-	if err != nil {
-		return err
-	}
-	config, err := svc.AppConfig(ctx)
-	if err != nil {
-		return err
-	}
-
-	changeEmail := fleet.Email{
-		Subject: "Confirm Fleet Email Change",
-		To:      []string{email},
-		Config:  config,
-		Mailer: &mail.ChangeEmailMailer{
-			Token:    token,
-			BaseURL:  template.URL(config.ServerSettings.ServerURL + svc.config.Server.URLPrefix),
-			AssetURL: getAssetURL(),
-		},
-	}
-	return svc.mailService.SendEmail(changeEmail)
-}
-
-func (svc *Service) DeleteUser(ctx context.Context, id uint) error {
-	if err := svc.authz.Authorize(ctx, &fleet.User{ID: id}, fleet.ActionWrite); err != nil {
-		return err
-	}
-
-	return svc.ds.DeleteUser(ctx, id)
-}
-
 func (svc *Service) ChangeUserEmail(ctx context.Context, token string) (string, error) {
 	vc, ok := viewer.FromContext(ctx)
 	if !ok {
@@ -268,14 +95,6 @@ func (svc *Service) ChangeUserEmail(ctx context.Context, token string) (string, 
 	}
 
 	return svc.ds.ConfirmPendingEmailChange(ctx, vc.UserID(), token)
-}
-
-func (svc *Service) User(ctx context.Context, id uint) (*fleet.User, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.User{ID: id}, fleet.ActionRead); err != nil {
-		return nil, err
-	}
-
-	return svc.ds.UserByID(ctx, id)
 }
 
 func (svc *Service) UserUnauthorized(ctx context.Context, id uint) (*fleet.User, error) {
@@ -299,14 +118,6 @@ func (svc *Service) AuthenticatedUser(ctx context.Context) (*fleet.User, error) 
 	return vc.User, nil
 }
 
-func (svc *Service) ListUsers(ctx context.Context, opt fleet.UserListOptions) ([]*fleet.User, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.User{}, fleet.ActionRead); err != nil {
-		return nil, err
-	}
-
-	return svc.ds.ListUsers(ctx, opt)
-}
-
 // setNewPassword is a helper for changing a user's password. It should be
 // called to set the new password after proper authorization has been
 // performed.
@@ -323,33 +134,6 @@ func (svc *Service) setNewPassword(ctx context.Context, user *fleet.User, passwo
 		return ctxerr.Wrap(ctx, err, "saving changed password")
 	}
 
-	return nil
-}
-
-func (svc *Service) ChangePassword(ctx context.Context, oldPass, newPass string) error {
-	vc, ok := viewer.FromContext(ctx)
-	if !ok {
-		return fleet.ErrNoContext
-	}
-
-	if err := svc.authz.Authorize(ctx, vc.User, fleet.ActionWrite); err != nil {
-		return err
-	}
-
-	if vc.User.SSOEnabled {
-		return ctxerr.New(ctx, "change password for single sign on user not allowed")
-	}
-	if err := vc.User.ValidatePassword(newPass); err == nil {
-		return fleet.NewInvalidArgumentError("new_password", "cannot reuse old password")
-	}
-
-	if err := vc.User.ValidatePassword(oldPass); err != nil {
-		return ctxerr.Wrap(ctx, fleet.NewInvalidArgumentError("old_password", "old password does not match"))
-	}
-
-	if err := svc.setNewPassword(ctx, vc.User, newPass); err != nil {
-		return ctxerr.Wrap(ctx, err, "setting new password")
-	}
 	return nil
 }
 
@@ -431,34 +215,6 @@ func (svc *Service) PerformRequiredPasswordReset(ctx context.Context, password s
 	return user, nil
 }
 
-func (svc *Service) RequirePasswordReset(ctx context.Context, uid uint, require bool) (*fleet.User, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.User{ID: uid}, fleet.ActionWrite); err != nil {
-		return nil, err
-	}
-
-	user, err := svc.ds.UserByID(ctx, uid)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "loading user by ID")
-	}
-	if user.SSOEnabled {
-		return nil, ctxerr.New(ctx, "password reset for single sign on user not allowed")
-	}
-	// Require reset on next login
-	user.AdminForcedPasswordReset = require
-	if err := svc.saveUser(ctx, user); err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "saving user")
-	}
-
-	if require {
-		// Clear all of the existing sessions
-		if err := svc.DeleteSessionsForUser(ctx, user.ID); err != nil {
-			return nil, ctxerr.Wrap(ctx, err, "deleting user sessions")
-		}
-	}
-
-	return user, nil
-}
-
 func (svc *Service) RequestPasswordReset(ctx context.Context, email string) error {
 	// skipauth: No viewer context available. The user is locked out of their
 	// account and trying to reset their password.
@@ -511,11 +267,4 @@ func (svc *Service) RequestPasswordReset(ctx context.Context, email string) erro
 	}
 
 	return svc.mailService.SendEmail(resetEmail)
-}
-
-// saves user in datastore.
-// doesn't need to be exposed to the transport
-// the service should expose actions for modifying a user instead
-func (svc *Service) saveUser(ctx context.Context, user *fleet.User) error {
-	return svc.ds.SaveUser(ctx, user)
 }
