+++ server/service/sessions_test.go
@@ -1,0 +1,8 @@
+package service
+
+// TODO(mna): when migrating Session-related endpoints, add auth tests for those
+// endpoints (the auth is session-based, not user-based).
+//_, err = svc.GetInfoAboutSessionsForUser(ctx, 999)
+//checkAuthErr(t, tt.shouldFailTeamWrite, err)
+//_, err = svc.GetInfoAboutSessionsForUser(ctx, 888)
+//checkAuthErr(t, tt.shouldFailGlobalWrite, err)
